refactor: move camera and pixel wrapper to render package
